geolocation-API/watchPosition_permission_deny.https.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Geolocation Test: watchPosition location access denied</title>
<link rel="help" href="http://www.w3.org/TR/geolocation-API/#privacy_for_uas" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
  promise_test(async (t) => {
    await test_driver.set_permission({ name: "geolocation" }, "denied");
    const promise = new Promise((resolve, reject) => {
      navigator.geolocation.watchPosition(reject, resolve);
    });
    const result = await promise;
    assert_true(
      result instanceof GeolocationPositionError,
      "expected GeolocationPositionError"
    );
    await test_driver.set_permission({ name: "geolocation" }, "prompt");
  }, "User denies access, check that error callback is called.");
</script>
